Separate TrayIcon class and create activity window
